+++ src/idunno.Bluesky/Embed/OpenGraphEmbeddedCardGenerator.cs
@@ -42,6 +42,7 @@ public partial class OpenGraphEmbeddedCardGenerator : BaseEmbeddedCardGenerator
         : base(agent, agent?.HttpClient, logger)
     {
         ArgumentNullException.ThrowIfNull(agent);
+        ArgumentNullException.ThrowIfNull(agent.HttpClient);
     }
 
     /// <summary>
@@ -80,8 +81,14 @@ public partial class OpenGraphEmbeddedCardGenerator : BaseEmbeddedCardGenerator
         return null;
     }
 
+    /// <summary>Creates an <see cref="EmbeddedExternal"/> from OpenGraph metadata.</summary>
+    /// <param name="uri">The URI to retrieve OpenGraph data from.</param>
+    /// <param name="pageContent">The HTML content of the page.</param>
+    /// <param name="cancellationToken">A token to cancel the operation.</param>
+    /// <returns>An <see cref="EmbeddedExternal"/> if OpenGraph data is found; otherwise, <see langword="null"/>.</returns>
+    /// <exception cref="ArgumentNullException">Thrown if <paramref name="uri"/> is <see langword="null" /></exception>
     [SuppressMessage("Minor Code Smell", "S3267:Loops should be simplified with \"LINQ\" expressions", Justification = "Avoid linq allocations in a hot path.")]
-    private async Task<EmbeddedExternal?> CreateEmbeddedExternalFromOpenGraphMetadata(Uri uri, string pageContent, CancellationToken cancellationToken = default)
+    protected async Task<EmbeddedExternal?> CreateEmbeddedExternalFromOpenGraphMetadata(Uri uri, string pageContent, CancellationToken cancellationToken = default)
     {
         ArgumentNullException.ThrowIfNull(uri);